+++ buildSrc/build.gradle.kts
@@ -5,3 +5,11 @@ plugins {
 repositories {
     gradlePluginPortal()
 }
+
+dependencies {
+    implementation("com.github.spotbugs.snom:spotbugs-gradle-plugin:5.2.5")
+    implementation("org.sonarsource.scanner.gradle:sonarqube-gradle-plugin:4.4.1.3373")
+    implementation("net.ltgt.gradle:gradle-errorprone-plugin:3.1.0")
+    implementation("info.solidsoft.gradle.pitest:gradle-pitest-plugin:1.15.0")
+    implementation("org.gradle:test-retry-gradle-plugin:1.5.7")
+}
