+++ pg-index-health-testing/build.gradle.kts
@@ -1,3 +1,8 @@
+plugins {
+    id("pg-index-health.java-conventions")
+    id("pg-index-health.publish")
+}
+
 description = "pg-index-health-testing is an auxiliary library that allows you to run a PostgreSQL cluster in tests."
 
 dependencies {
